# api/openapi/mono.yaml
openapi: 3.0.3
info:
  title: Mono Task Management API
  version: 1.0.0
  description: RESTful API for task and list management with recurring templates
servers:
  - url: http://localhost:8081/api
    description: Local development
security:
  - BearerAuth: []
paths:
...
  /v1/lists/{list_id}/recurring-templates:
    post:
      operationId: createRecurringTemplate
      summary: Create a recurring task template
      tags: [RecurringTemplates]
      parameters:
        - name: list_id
          in: path
          required: true
          description: List ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRecurringTemplateRequest'
      responses:
        '201':
          description: Template created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateRecurringTemplateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

    get:
      operationId: listRecurringTemplates
      summary: List recurring templates for a list
      tags: [RecurringTemplates]
      parameters:
        - name: list_id
          in: path
          required: true
          description: List ID
          schema:
            type: string
            format: uuid
        - name: active_only
          in: query
          description: Filter for active templates only
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Templates retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListRecurringTemplatesResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/recurring-templates/{id}:
    get:
      operationId: getRecurringTemplate
      summary: Get a recurring template by ID
      tags: [RecurringTemplates]
      parameters:
        - name: id
          in: path
          required: true
          description: Template ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Template found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetRecurringTemplateResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

    patch:
      operationId: updateRecurringTemplate
      summary: Update a recurring template
      tags: [RecurringTemplates]
      parameters:
        - name: id
          in: path
          required: true
          description: Template ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRecurringTemplateRequest'
      responses:
        '200':
          description: Template updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateRecurringTemplateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      operationId: deleteRecurringTemplate
      summary: Delete a recurring template
      tags: [RecurringTemplates]
      parameters:
        - name: id
          in: path
          required: true
          description: Template ID
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Template deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  schemas:
...
    CreateRecurringTemplateRequest:
      type: object
      required:
        - title
        - recurrence_pattern
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        tags:
          type: array
          items:
            type: string
        priority:
          $ref: '#/components/schemas/TaskPriority'
        estimated_duration:
          type: string
          description: ISO 8601 duration
        recurrence_pattern:
          $ref: '#/components/schemas/RecurrencePattern'
        recurrence_config:
          type: string
          description: JSON config for pattern-specific settings
        due_offset:
          type: string
          description: ISO 8601 duration offset from instance date
        generation_window_days:
          type: integer
          minimum: 1
          maximum: 365
          default: 30
    CreateRecurringTemplateResponse:
      type: object
      properties:
        template:
          $ref: '#/components/schemas/RecurringTaskTemplate'
    GetRecurringTemplateResponse:
      type: object
      properties:
        template:
          $ref: '#/components/schemas/RecurringTaskTemplate'
    UpdateRecurringTemplateRequest:
      type: object
      properties:
        template:
          $ref: '#/components/schemas/RecurringTaskTemplate'
        update_mask:
          type: array
          items:
            type: string
          description: Fields to update (field mask)
    UpdateRecurringTemplateResponse:
      type: object
      properties:
        template:
          $ref: '#/components/schemas/RecurringTaskTemplate'
    ListRecurringTemplatesResponse:
      type: object
      properties:
        templates:
          type: array
          items:
            $ref: '#/components/schemas/RecurringTaskTemplate'

    # Domain models
...
    RecurringTaskTemplate:
      type: object
      properties:
        id:
          type: string
          format: uuid
        list_id:
          type: string
          format: uuid
        title:
          type: string
        tags:
          type: array
          items:
            type: string
        priority:
          $ref: '#/components/schemas/TaskPriority'
        estimated_duration:
          type: string
          description: ISO 8601 duration
        recurrence_pattern:
          $ref: '#/components/schemas/RecurrencePattern'
        recurrence_config:
          type: string
          description: JSON configuration
        due_offset:
          type: string
          description: ISO 8601 duration
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        last_generated_until:
          type: string
          format: date-time
        generation_window_days:
          type: integer

    # Enums
...
    TaskPriority:
      type: string
      enum:
        - low
        - medium
        - high
        - urgent
    RecurrencePattern:
      type: string
      enum:
        - daily
        - weekly
        - biweekly
        - monthly
        - yearly
        - quarterly
        - weekdays

    # Error schemas
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  issue:
                    type: string
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized - invalid or missing API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: API key passed as Bearer token in Authorization header